Implement dynamic import for OneSignal
* Introduced a wrapper module for react-onesignal to handle dynamic
imports and ensure compatibility with ES module changes.
* Created a loadOneSignal function that dynamically imports
react-onesignal and caches the module for future use.
* This refactor addresses issues related to module type conflicts and
improves the maintainability of the codebase by centralizing the import
logic.
* Added a placeholder for additional utility functions to extend the
wrapper module's functionality as needed.

This is needed because react-onesignal v3.2.0 makes it a module without
making the index file a .mjs, which gives the following error:

ReferenceError: module is not defined in ES module scope
This file is being treated as an ES module because it has a '.js' file
extension and 'D:\Documents\repositories\asvgay\the-rhapsodies
\node_modules\react-onesignal\package.json' contains "type": "module".
To treat it as a CommonJS script, rename it to use the '.cjs' file
extension.

Signed-off-by: F.S. Koulen <feryll@hotmail.com>